fix(package partial): repos without documentation should work too
Earlier there were two separate partials, one to show the
packages without documentation, and ones with. I have now moved
this data to the JSON and a single partial uses that information
to point to the correct link.

This should solve the 404 that was being caused on the Browser
Search blog post.